+++ packages/vexide-devices/src/controller.rs
@@ -222,16 +222,19 @@ impl Future for ControllerScreenWriteFuture<'_> {
             column,
             text,
             controller,
-            enforce_visible: visible,
+            enforce_visible,
         } = state
         {
+            if *enforce_visible {
+                assert!(
+                    (*line != 0 && *line <= ControllerScreen::MAX_LINES as u8),
+                    "Invalid line number ({line}) is greater than the maximum number of lines ({})",
+                    ControllerScreen::MAX_LINES
+                );
+            }
+
             assert!(
-                *visible && (*line == 0 || *line > ControllerScreen::MAX_LINES as u8),
-                "Invalid line number ({line}) is greater than the maximum number of lines ({})",
-                ControllerScreen::MAX_LINES
-            );
-            assert!(
-                *column == 0 || *column > ControllerScreen::MAX_COLUMNS as u8,
+                *column != 0 && *column <= ControllerScreen::MAX_COLUMNS as u8,
                 "Invalid column number ({column}) is greater than the maximum number of columns ({})",
                 ControllerScreen::MAX_COLUMNS
             );
@@ -536,14 +539,14 @@ impl ControllerScreen {
         validate_connection(self.id)?;
 
         assert!(
-            column < Self::MAX_COLUMNS as u8 && column != 0,
+            column <= Self::MAX_COLUMNS as u8 && column != 0,
             "Invalid column number ({column}) is greater than the maximum number of columns ({})",
             ControllerScreen::MAX_COLUMNS
         );
         assert!(
-            line < Self::MAX_LINES as u8 && line != 0,
-            "Invalid column number ({column}) is greater than the maximum number of columns ({})",
-            ControllerScreen::MAX_COLUMNS
+            line <= Self::MAX_LINES as u8 && line != 0,
+            "Invalid line number ({line}) is greater than the maximum number of line ({})",
+            ControllerScreen::MAX_LINES
         );
 
         let id: V5_ControllerId = self.id.into();
@@ -980,7 +983,7 @@ impl Controller {
     /// }
     /// ```
     pub fn try_rumble(&mut self, pattern: impl AsRef<str>) -> Result<(), ControllerError> {
-        self.screen.try_set_text(pattern, 3, 0)
+        self.screen.try_set_text(pattern, 4, 1)
     }
 }